SCD30 improvements
The interface specification for the SCD30 requires up to 150ms of clock
stretching, but the default ESP8266 software implementation only allows
much less.  On ESP8266, increase it so that the sensor doesn't get
random faults.

The start measurement command requires a pressure compensation
parameter for it to have any effect, so supply one.  This is a
persistent setting, so it probably wasn't noticed before because
the saved value was silently used.

SCD30 sensors from the factory will not have automatic self calibration
enabled, so add an option (defaulting to on) to enable it.  This is
also important since the sensor will read nonsense until at least
one calibration has been completed so having a way to enable it is
a very good idea.

Add optional altitude compensation setting.  Since the start
measurement command overrides any that was set, add a configuration
setting so it can be set to the local conditions.